# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/announcements": {
      "get": {
        "description": "Public: active announcements only; admins see all.",
        "operationId": "AnnouncementsController_list",
        "parameters": [
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 500,
              "default": 100,
              "type": "number"
            }
          },
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "description": "Rows to skip, for paging",
            "schema": {
              "minimum": 0,
              "default": 0,
              "type": "number"
            }
          },
          {
            "name": "order",
            "required": false,
            "in": "query",
            "schema": {
              "default": "-createdAt",
              "type": "string",
              "enum": [
                "id",
                "-id",
                "title",
                "-title",
                "body",
                "-body",
                "date",
                "-date",
                "type",
                "-type",
                "status",
                "-status",
                "createdAt",
                "-createdAt",
                "updatedAt",
                "-updatedAt",
                "titleHi",
                "-titleHi",
                "bodyHi",
                "-bodyHi"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AnnouncementVo"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid order, limit or filter",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorVo"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List announcements",
        "tags": [
          "announcements"
        ]
      },
      "post": {
        "description": "Admin only.",
        "operationId": "AnnouncementsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAnnouncementDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnnouncementVo"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorVo"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorVo"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorVo"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a announcement",
        "tags": [
          "announcements"
        ]
      }
    },
    "/api/v1/announcements/{id}": {
      "patch": {
        "description": "Admin only.",
        "operationId": "AnnouncementsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Record id",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAnnouncementDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnnouncementVo"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorVo"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorVo"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorVo"
                }
              }
            }
          },
          "404": {
            "description": "Record not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorVo"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update a announcement",
        "tags": [
          "announcements"
        ]
      },
      "delete": {
        "description": "Admin only.",
        "operationId": "AnnouncementsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Record id",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Deleted (also when the id did not exist)"
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorVo"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorVo"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete a announcement",
        "tags": [
          "announcements"
        ]
      }
    },
...
    "/api/v1/events": {
      "get": {
        "description": "Public: published events only; admins see all.",
        "operationId": "EventsController_list",
        "parameters": [
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 500,
              "default": 100,
              "type": "number"
            }
          },
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "description": "Rows to skip, for paging",
            "schema": {
              "minimum": 0,
              "default": 0,
              "type": "number"
            }
          },
          {
            "name": "order",
            "required": false,
            "in": "query",
            "schema": {
              "default": "-createdAt",
              "type": "string",
              "enum": [
                "id",
                "-id",
                "title",
                "-title",
                "slug",
                "-slug",
                "bannerUrl",
                "-bannerUrl",
                "description",
                "-description",
                "date",
                "-date",
                "startTime",
                "-startTime",
                "endTime",
                "-endTime",
                "venue",
                "-venue",
                "mapLocation",
                "-mapLocation",
                "organizer",
                "-organizer",
                "contact",
                "-contact",
                "registrationOpen",
                "-registrationOpen",
                "registrationClose",
                "-registrationClose",
                "fee",
                "-fee",
                "capacity",
                "-capacity",
                "rules",
                "-rules",
                "terms",
                "-terms",
                "status",
                "-status",
                "createdAt",
                "-createdAt",
                "updatedAt",
                "-updatedAt",
                "titleHi",
                "-titleHi",
                "descriptionHi",
                "-descriptionHi"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EventVo"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid order, limit or filter",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorVo"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List events",
        "tags": [
          "events"
        ]
      },
      "post": {
        "description": "Admin only.",
        "operationId": "EventsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateEventDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventVo"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorVo"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorVo"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorVo"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a event",
        "tags": [
          "events"
        ]
      }
    },
    "/api/v1/events/{id}": {
      "patch": {
        "description": "Admin only.",
        "operationId": "EventsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Record id",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateEventDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventVo"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorVo"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorVo"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorVo"
                }
              }
            }
          },
          "404": {
            "description": "Record not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorVo"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update a event",
        "tags": [
          "events"
        ]
      },
      "delete": {
        "description": "Admin only.",
        "operationId": "EventsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Record id",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Deleted (also when the id did not exist)"
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorVo"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorVo"
                }
              }
            }
          },
          "409": {
            "description": "The event has registrations",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorVo"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete a event",
        "tags": [
          "events"
        ]
      }
    },
...
  },
  "info": {
    "title": "NPS Indore portal API",
    "description": "Backend API for the NPS Indore member portal. Errors are always `{ \"error\": \"message\" }`.",
    "version": "1",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Session token (also sent as the httpOnly npsi_session cookie)"
      }
    },
    "schemas": {
...
      "ErrorVo": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "example": "Authentication required."
          }
        },
        "required": [
          "error"
        ]
      },
...
      "AnnouncementVo": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "body": {
            "type": "string"
          },
          "date": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "type": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "titleHi": {
            "type": "string",
            "nullable": true
          },
          "bodyHi": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "id",
          "title",
          "body",
          "date",
          "type",
          "status",
          "createdAt",
          "updatedAt",
          "titleHi",
          "bodyHi"
        ]
      },
      "CreateAnnouncementDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "body": {
            "type": "string"
          },
          "date": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "type": {
            "type": "string",
            "enum": [
              "General",
              "Event",
              "Urgent",
              "Festival"
            ],
            "nullable": true
          },
          "status": {
            "type": "string",
            "enum": [
              "Active",
              "Archived"
            ],
            "nullable": true
          },
          "titleHi": {
            "type": "string",
            "nullable": true
          },
          "bodyHi": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "title",
          "body"
        ]
      },
      "UpdateAnnouncementDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "body": {
            "type": "string"
          },
          "date": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "type": {
            "type": "string",
            "enum": [
              "General",
              "Event",
              "Urgent",
              "Festival"
            ],
            "nullable": true
          },
          "status": {
            "type": "string",
            "enum": [
              "Active",
              "Archived"
            ],
            "nullable": true
          },
          "titleHi": {
            "type": "string",
            "nullable": true
          },
          "bodyHi": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
      "EventVo": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "slug": {
            "type": "string",
            "nullable": true
          },
          "bannerUrl": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "date": {
            "type": "string",
            "format": "date"
          },
          "startTime": {
            "type": "string",
            "nullable": true
          },
          "endTime": {
            "type": "string",
            "nullable": true
          },
          "venue": {
            "type": "string"
          },
          "mapLocation": {
            "type": "string",
            "nullable": true
          },
          "organizer": {
            "type": "string",
            "nullable": true
          },
          "contact": {
            "type": "string",
            "nullable": true
          },
          "registrationOpen": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "registrationClose": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "fee": {
            "type": "number",
            "nullable": true
          },
          "capacity": {
            "type": "integer",
            "nullable": true
          },
          "rules": {
            "type": "string",
            "nullable": true
          },
          "terms": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "titleHi": {
            "type": "string",
            "nullable": true
          },
          "descriptionHi": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "id",
          "title",
          "slug",
          "bannerUrl",
          "description",
          "date",
          "startTime",
          "endTime",
          "venue",
          "mapLocation",
          "organizer",
          "contact",
          "registrationOpen",
          "registrationClose",
          "fee",
          "capacity",
          "rules",
          "terms",
          "status",
          "createdAt",
          "updatedAt",
          "titleHi",
          "descriptionHi"
        ]
      },
      "CreateEventDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "slug": {
            "type": "string",
            "nullable": true
          },
          "bannerUrl": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "date": {
            "type": "string",
            "format": "date-time"
          },
          "startTime": {
            "type": "string",
            "nullable": true
          },
          "endTime": {
            "type": "string",
            "nullable": true
          },
          "venue": {
            "type": "string"
          },
          "mapLocation": {
            "type": "string",
            "nullable": true
          },
          "organizer": {
            "type": "string",
            "nullable": true
          },
          "contact": {
            "type": "string",
            "nullable": true
          },
          "registrationOpen": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "registrationClose": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "fee": {
            "type": "number",
            "nullable": true
          },
          "capacity": {
            "type": "integer",
            "nullable": true
          },
          "rules": {
            "type": "string",
            "nullable": true
          },
          "terms": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "string",
            "enum": [
              "DRAFT",
              "PUBLISHED",
              "ARCHIVED"
            ],
            "nullable": true
          },
          "titleHi": {
            "type": "string",
            "nullable": true
          },
          "descriptionHi": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "title",
          "date",
          "venue"
        ]
      },
      "UpdateEventDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "slug": {
            "type": "string",
            "nullable": true
          },
          "bannerUrl": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "date": {
            "type": "string",
            "format": "date-time"
          },
          "startTime": {
            "type": "string",
            "nullable": true
          },
          "endTime": {
            "type": "string",
            "nullable": true
          },
          "venue": {
            "type": "string"
          },
          "mapLocation": {
            "type": "string",
            "nullable": true
          },
          "organizer": {
            "type": "string",
            "nullable": true
          },
          "contact": {
            "type": "string",
            "nullable": true
          },
          "registrationOpen": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "registrationClose": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "fee": {
            "type": "number",
            "nullable": true
          },
          "capacity": {
            "type": "integer",
            "nullable": true
          },
          "rules": {
            "type": "string",
            "nullable": true
          },
          "terms": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "string",
            "enum": [
              "DRAFT",
              "PUBLISHED",
              "ARCHIVED"
            ],
            "nullable": true
          },
          "titleHi": {
            "type": "string",
            "nullable": true
          },
          "descriptionHi": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
    }
  }
}